requirements.txt
python-bitcoinrpc>=1.0
flask>=0.12.2
pymongo>=3.4.0
